feat(stats): cleanup
