+++ src/app/u/[username]/stats/[year]/page.tsx
@@ -1,6 +1,5 @@
 import { Suspense } from 'react';
 
-import { cx } from 'class-variance-authority';
 import Image from 'next/image';
 import Link from 'next/link';
 import { notFound } from 'next/navigation';
@@ -9,9 +8,11 @@ import Page from '@/components/Page/Page';
 import SkeletonList from '@/components/Skeletons/SkeletonList';
 import SpotlightBackground from '@/components/Spotlight/SpotlightBackground';
 import SpotlightTitle from '@/components/Spotlight/SpotlightTitle';
+import Block from '@/components/Stats/Block';
 import MostWatchedGenres from '@/components/Stats/MostWatchedGenres';
 import MostWatchedProviders from '@/components/Stats/MostWatchedProviders';
 import PopularNotWatched from '@/components/Stats/PopularNotWatched';
+import SvgPattern from '@/components/Stats/SvgPattern';
 import WatchedByYear from '@/components/Stats/Watched';
 import WatchedPerWeek from '@/components/Stats/WatchedPerWeek';
 import WorldMap from '@/components/Stats/WorldMap';
@@ -21,100 +22,6 @@ import { fetchTvSeries } from '@/lib/tmdb';
 type Props = Readonly<{
   params: Promise<{ username: string; year: number }>;
 }>;
-
-const colours = ['#666666', '#00FFFF', '#FF0080']; // darker grey, cyan, magenta
-
-const SvgPattern = ({ className }: Readonly<{ className?: string }>) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    viewBox="0 0 480 390"
-    className={cx('scale-y-150 md:scale-y-125 lg:scale-y-100', className)}
-  >
-    <g fill="none" fillRule="evenodd">
-      <g transform="translate(447 121)">
-        <rect fill={colours[0]} y="55" width="5" height="90"></rect>
-        <rect fill={colours[1]} width="5" height="48"></rect>
-        <rect fill={colours[2]} y="169" width="5" height="22"></rect>
-      </g>
-      <g transform="translate(417 111)">
-        <rect fill={colours[0]} y="66" width="5" height="80"></rect>
-        <rect fill={colours[1]} width="5" height="38"></rect>
-        <rect fill={colours[2]} y="152" width="5" height="42"></rect>
-      </g>
-      <g transform="translate(387 84)">
-        <rect fill={colours[0]} y="78" width="5" height="100"></rect>
-        <rect fill={colours[2]} y="201" width="5" height="42"></rect>
-        <rect fill={colours[1]} width="5" height="70"></rect>
-      </g>
-      <g transform="translate(357 80)">
-        <rect fill={colours[0]} y="96" width="5" height="85"></rect>
-        <rect fill={colours[0]} y="68" width="5" height="20"></rect>
-        <rect fill={colours[1]} width="5" height="57"></rect>
-        <rect fill={colours[2]} y="212" width="5" height="42"></rect>
-      </g>
-      <g transform="translate(327 96)">
-        <rect fill={colours[0]} y="48" width="5" height="80"></rect>
-        <rect fill={colours[1]} y="12" width="5" height="28"></rect>
-        <rect fill={colours[1]} width="5" height="4"></rect>
-        <rect fill={colours[2]} y="144" width="5" height="32"></rect>
-      </g>
-      <g transform="translate(297 111)">
-        <rect fill={colours[0]} y="46" width="5" height="70"></rect>
-        <rect fill={colours[0]} y="124" width="5" height="4"></rect>
-        <rect fill={colours[1]} width="5" height="38"></rect>
-        <rect fill={colours[2]} y="142" width="5" height="32"></rect>
-      </g>
-      <g transform="translate(267 40)">
-        <rect fill={colours[0]} y="102" width="5" height="90"></rect>
-        <rect fill={colours[2]} y="210" width="5" height="22"></rect>
-        <rect fill={colours[1]} width="5" height="80"></rect>
-      </g>
-      <g transform="translate(237 86)">
-        <rect fill={colours[0]} y="36" width="5" height="90"></rect>
-        <rect fill={colours[1]} width="5" height="28"></rect>
-        <rect fill={colours[2]} y="132" width="5" height="42"></rect>
-      </g>
-      <g transform="translate(207 86)">
-        <rect fill={colours[0]} y="36" width="5" height="70"></rect>
-        <rect fill={colours[1]} width="5" height="28"></rect>
-        <rect fill={colours[2]} y="132" width="5" height="32"></rect>
-      </g>
-      <g transform="translate(177 111)">
-        <rect fill={colours[0]} y="36" width="5" height="85"></rect>
-        <rect fill={colours[1]} width="5" height="28"></rect>
-        <rect fill={colours[2]} y="132" width="5" height="42"></rect>
-      </g>
-      <g transform="translate(147 82)">
-        <rect fill={colours[0]} y="102" width="5" height="90"></rect>
-        <rect fill={colours[2]} y="210" width="5" height="22"></rect>
-        <rect fill={colours[1]} width="5" height="80"></rect>
-      </g>
-      <g transform="translate(117 74)">
-        <rect fill={colours[0]} y="51" width="5" height="160"></rect>
-        <rect fill={colours[1]} width="5" height="14"></rect>
-        <rect fill={colours[1]} y="23" width="5" height="20"></rect>
-        <rect fill={colours[2]} y="246" width="5" height="24"></rect>
-      </g>
-      <g transform="translate(87 127)">
-        <rect fill={colours[0]} y="64" width="5" height="50"></rect>
-        <rect fill={colours[2]} y="144" width="5" height="64"></rect>
-        <rect fill={colours[1]} width="5" height="30"></rect>
-      </g>
-      <g transform="translate(57 77)">
-        <rect fill={colours[0]} y="104" width="5" height="80"></rect>
-        <rect fill={colours[2]} y="204" width="5" height="74"></rect>
-        <rect fill={colours[1]} width="5" height="34"></rect>
-        <rect fill={colours[1]} y="32" width="5" height="40"></rect>
-      </g>
-      <g transform="translate(27 117)">
-        <rect fill={colours[0]} y="84" width="5" height="80"></rect>
-        <rect fill={colours[2]} y="184" width="5" height="34"></rect>
-        <rect fill={colours[2]} y="232" width="5" height="14"></rect>
-        <rect fill={colours[1]} width="5" height="40"></rect>
-      </g>
-    </g>
-  </svg>
-);
 
 export async function generateMetadata({ params }: Props) {
   const { username } = await params;
@@ -135,27 +42,6 @@ export default async function StatsByYearPage({ params }: Props) {
     return notFound();
   }
 
-  /*
-    - total days/hours watched
-    - total episodes watched
-    - total shows finished
-    - total shows unfinished
-    - total added to watchlist
-    - total favourited
-
-    - first play of 2024
-    - last play of 2024
-
-    - most watched genres
-    - most watched streaming services
-    - most watched countries (world map)
-    - most watched cast
-    - most watched directors/creators
-
-    - all watched shows in 2024 (grid)
-    - popular shows in 2024 you didn't watch
-  */
-
   const firstShow = await fetchTvSeries(66276, {
     includeImages: true,
   });
@@ -171,64 +57,26 @@ export default async function StatsByYearPage({ params }: Props) {
           <SvgPattern className="absolute left-[1rem] top-0 w-[280px] sm:w-[320px] md:w-[400px] lg:-top-8 lg:w-[480px]" />
           <SvgPattern className="absolute right-[1rem] top-0 w-[280px] scale-x-[-1] sm:w-[320px] md:w-[400px] lg:-top-8 lg:w-[480px]" />
         </div>
-
         <div className="absolute inset-0 flex flex-col items-center justify-center">
           <h1 className="text-[6rem] font-extrabold leading-none tracking-wide sm:text-[8rem] md:text-[10rem]">
-            2024
+            {year}
           </h1>
           <span className="mt-2 text-sm text-white/55 md:text-lg">
             A year of tvseri.es with
-            <span
-              className="ml-2 rounded px-2 py-1 text-white"
-              style={{
-                backgroundColor: colours[0],
-              }}
-            >
-              koenoe
+            <span className="ml-2 rounded bg-[#666] px-2 py-1 text-white">
+              {user.username}
             </span>
           </span>
         </div>
       </div>
       <div className="container mt-10 md:mt-8">
         <div className="grid grid-cols-2 gap-4 md:grid-cols-3 xl:grid-cols-6">
-          <div className="flex flex-col justify-center gap-y-1 whitespace-nowrap rounded bg-[#333] p-3 md:gap-y-2 md:p-6">
-            <span className="text-[0.65rem] text-white/60 md:text-sm">
-              Total runtime
-            </span>
-            <span className="text-lg font-semibold md:text-2xl">
-              17d 15h 59m
-            </span>
-          </div>
-          <div className="flex flex-col justify-center gap-y-1 whitespace-nowrap rounded bg-[#333] p-3 md:gap-y-2 md:p-6">
-            <span className="text-[0.65rem] text-white/60 md:text-sm">
-              Episodes watched
-            </span>
-            <span className="text-lg font-semibold md:text-2xl">1,831</span>
-          </div>
-          <div className="flex flex-col justify-center gap-y-1 whitespace-nowrap rounded bg-[#333] p-3 md:gap-y-2 md:p-6">
-            <span className="text-[0.65rem] text-white/60 md:text-sm">
-              Series finished
-            </span>
-            <span className="text-lg font-semibold md:text-2xl">12</span>
-          </div>
-          <div className="flex flex-col justify-center gap-y-1 whitespace-nowrap rounded bg-[#333] p-3 md:gap-y-2 md:p-6">
-            <span className="text-[0.65rem] text-white/60 md:text-sm">
-              In progress
-            </span>
-            <span className="text-lg font-semibold md:text-2xl">18</span>
-          </div>
-          <div className="flex flex-col justify-center gap-y-1 whitespace-nowrap rounded bg-[#333] p-3 md:gap-y-2 md:p-6">
-            <span className="text-[0.65rem] text-white/60 md:text-sm">
-              Added to favorites
-            </span>
-            <span className="text-lg font-semibold md:text-2xl">78</span>
-          </div>
-          <div className="flex flex-col justify-center gap-y-1 whitespace-nowrap rounded bg-[#333] p-3 md:gap-y-2 md:p-6">
-            <span className="text-[0.65rem] text-white/60 md:text-sm">
-              Want to watch
-            </span>
-            <span className="text-lg font-semibold md:text-2xl">36</span>
-          </div>
+          <Block label="Total runtime" value="17d 15h 59m" />
+          <Block label="Episodes watched" value="1,831" />
+          <Block label="Series finished" value="12" />
+          <Block label="In progress" value="18" />
+          <Block label="Added to favorites" value="78" />
+          <Block label="Want to watch" value="36" />
         </div>
         <div className="relative mt-14 grid w-full grid-cols-1 gap-20 md:mt-20 xl:grid-cols-2 xl:gap-10">
           <div>
